requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pdfminer.six==20221105
scikit-learn==1.3.2
numpy==1.26.2
